Updated strlen function return datatype.

+++ ft_strlen.c
@@ -14,7 +14,7 @@
 
 // Code loops through a string of characters, for each character it
 // increments index, at the end once the loop reaches Null it returns index
-int ft_strlen(const char *str)
+size_t ft_strlen(const char *str)
 {
     size_t index;
